Create timeouts_controller and tests work
