+++ decidim-core/app/assets/javascripts/decidim/session_timeouter.js.es6
@@ -8,25 +8,22 @@
     const popup = new Foundation.Reveal($timeoutModal);
     const $continueSessionButton = $("#continueSession")
 
-    const resetTimer = (secondsUntilExpiration) => {
+    const setTimer = (secondsUntilExpiration) => {
+      if (!secondsUntilExpiration) {
+        return;
+      }
       endsAt = exports.moment().add(secondsUntilExpiration, "seconds")
     }
 
     const sessionTimeLeft = async () => {
-      $("#test_element").append("Asking how much time is left<br>");
       const result = await $.ajax({
         method: "GET",
         url: secondsUntilTimeoutPath,
         contentType: "application/json",
         headers: {
           "X-CSRF-Token": $("meta[name=csrf-token]").attr("content")
-        },
-        complete: (jqXHR, textStatus) => {
-          $("#test_element").append(`Request status: ${textStatus}<br>`);
         }
       });
-      $("#test_element").append("TIME LEFT ASKING DONE<br>");
-      $("#test_element").append(`${JSON.stringify(result)}<br>`);
       return result.seconds_remaining
     }
 
@@ -45,18 +42,13 @@
       const diff = endsAt - exports.moment();
       const diffInSeconds = Math.round(diff / 1000);
 
-      console.log("diffInSeconds", diffInSeconds)
-      $("#test_element").append(`DIFF IN SECONDS: ${diffInSeconds}<br>`);
-      $("#test_element").append("Hello from the exitInterval<br>");
       if (diffInSeconds <= 90) {
         $timeoutModal.find("#reveal-hidden-sign-out")[0].click();
       } else if (diffInSeconds <= 150) {
         const secondsUntilSessionExpires = await sessionTimeLeft();
-        console.log("secondsUntilSessionExpires", secondsUntilSessionExpires)
+        setTimer(secondsUntilSessionExpires)
         if (secondsUntilSessionExpires <= 150) {
           popup.open();
-        } else {
-          resetTimer(secondsUntilSessionExpires)
         }
       }
     }, 10000);
@@ -64,14 +56,14 @@
     // Devise restarts its own timer on ajax requests,
     // so here we restart our.
     $(document).on("ajax:complete", () => {
-      resetTimer(sessionTimeOutInSeconds);
+      setTimer(sessionTimeOutInSeconds);
     });
 
     $(document).ajaxComplete((_event, _xhr, settings) => {
-      if (settings && settings.url === "/session_seconds_until_timeout") {
+      if (settings && settings.url === secondsUntilTimeoutPath) {
         return;
       }
-      resetTimer(sessionTimeOutInSeconds);
+      setTimer(sessionTimeOutInSeconds);
     });
 
     window.addEventListener("beforeunload", () => {
