+++ decidim-core/config/routes.rb
@@ -27,8 +27,7 @@ Decidim::Core::Engine.routes.draw do
 
   devise_scope :user do
     post "omniauth_registrations" => "devise/omniauth_registrations#create"
-    post "session_heartbeat" => "devise/sessions#heartbeat"
-    get "session_seconds_until_timeout" => "devise/sessions#seconds_until_timeout"
+
   end
 
   resource :locale, only: [:create]
@@ -104,6 +103,11 @@ Decidim::Core::Engine.routes.draw do
     get "activity", to: "user_activities#index", as: "profile_activity"
     get "timeline", to: "user_timeline#index", as: "profile_timeline"
     resources :conversations, except: [:destroy], controller: "user_conversations", as: "profile_conversations"
+  end
+
+  scope :timeouts do
+    post "heartbeat" => "timeouts#heartbeat"
+    get "seconds_until_timeout" => "timeouts#seconds_until_timeout"
   end
 
   resources :pages, only: [:index, :show], format: false
