+++ decidim-core/spec/system/session_timeout_spec.rb
@@ -25,31 +25,21 @@ describe "Session timeout", type: :system do
     end
 
     it "doesnt timeout when user is active in another window" do
-      win1 = current_window
       Devise.timeout_in = 2.minutes
+      win1 = current_window
       visit decidim.root_path
-      puts "START"
-      win2 = open_new_window
-      switch_to_window(win2)
+
+      switch_to_window(open_new_window)
       visit decidim.root_path
-      puts "FIRST CLICK (20s)"
-      find(".logo-wrapper", wait: 20).click
-      puts "SECOND CLICK (20s)"
-      find(".logo-wrapper", wait: 20).click
-      puts "THIRD CLICK (20s)"
-      find(".logo-wrapper", wait: 20).click
+      expect(page).to have_content("If you continue being inactive", wait: 11)
+      find("#continueSession").click
+      expect(page).to have_content("If you continue being inactive", wait: 11)
+      find("#continueSession").click
+      expect(page).to have_content("If you continue being inactive", wait: 11)
+      find("#continueSession").click
 
       switch_to_window(win1)
-
-      puts "HAVE TEST CONTENT (11s = 20*3+11)"
-      expect(page).to have_content("DIFF IN SECONDS", wait: 11)
-      puts "71s =>"
-      puts find("#test_element").text.inspect
-      expect(page).to have_content("TIME LEFT ASKING DONE")
-      expect(page).to have_content("If you continue being inactive")
-      puts "WAITING FOR INACTIVITY (20s)"
-      expect(page).not_to have_content("You were inactive for too long", wait: 20)
-      puts "91s =>"
+      expect(page).not_to have_content("You were inactive for too long")
     end
   end
 end
